decode: add DecodeImmutablesOnly() option
For performing validation on immutablity it is required to have
previous version of deserialized data. In case when we have only
previous version of the object it is suboptimal to enforce
deserialization of whole data structure when actually compering
versus immutable data is required.

This option adds possibility to decode only those fields which
are required for future compararison.